requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
npm~=0.1.1
